requirements.txt
streamlit>=1.30
requests>=2.31
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
google-generativeai